Create directories only if not already present in Linux client installer

+++ install_client_linux.sh
@@ -118,11 +118,11 @@ then
 	exit 2
 fi
 
-install -c -m 744 -d "$PREFIX/var/urbackup/data"
-install -c -m 744 -d "$PREFIX/sbin"
-install -c -m 744 -d "$PREFIX/bin"
-install -c -m 744 -d "$PREFIX/share/urbackup/scripts"
-install -c -m 744 -d "$PREFIX/etc/urbackup"
+test -e "$PREFIX/var/urbackup/data" || install -c -m 744 -d "$PREFIX/var/urbackup/data"
+test -e "$PREFIX/sbin" || install -c -m 744 -d "$PREFIX/sbin"
+test -e "$PREFIX/bin" || install -c -m 744 -d "$PREFIX/bin"
+test -e "$PREFIX/share/urbackup/scripts" || install -c -m 744 -d "$PREFIX/share/urbackup/scripts"
+test -e "$PREFIX/etc/urbackup" || install -c -m 744 -d "$PREFIX/etc/urbackup"
 
 install -c "$TARGET/urbackupclientbackend" "$PREFIX/sbin"
 install -c "$TARGET/urbackupclientctl" "$PREFIX/bin"
